requirements.txt
fastapi 
uvicorn 
spotipy 
python-dotenv 
abc 
typing 
os 
json 
requests 
logging 
urllib3 
uuid
base64
datetime
shutil
httpx
pydantic 
elevenlabs
python-multipart
sqlalchemy 
redis
celery
gevent #required to run celery library in windows
thefuzz
authlib 
pytest
fakeredis
unittest
re
unicodedata
starlette